refactor: reorganize controller package
