+++ src/main/scala/it/unibo/shipps/controller/BotTurnHandler.scala
@@ -2,6 +2,7 @@ package it.unibo.shipps.controller
 
 import it.unibo.shipps.model.player.Player
 import it.unibo.shipps.model.board.Position
+import it.unibo.shipps.model.{Turn, TurnLogic}
 import it.unibo.shipps.controller.GameStateManager.GameActionResult
 import it.unibo.shipps.controller.utils.DelayedExecutor
 import it.unibo.shipps.view.handler.TurnDialogHandler
