+++ src/main/scala/it/unibo/shipps/controller/positioning/PositioningController.scala
@@ -1,73 +1,13 @@
-package it.unibo.shipps.controller
+package it.unibo.shipps.controller.positioning
 
+import it.unibo.shipps.controller.GameState
 import it.unibo.shipps.model.*
 import it.unibo.shipps.model.board.{PlayerBoard, Position}
 import it.unibo.shipps.model.ship.Ship
+import it.unibo.shipps.model.{Turn, TurnLogic}
 
 /** Handles ship positioning logic and user interactions during positioning phase */
 object PositioningController:
-
-  /** Handles cell click during positioning phase
-    * @param gameState current game state
-    * @param position clicked position
-    * @param turn current turn
-    * @param positioning ship positioning logic
-    * @return updated game state
-    */
-  def handlePositioningClick(
-      gameState: GameState,
-      position: Position,
-      turn: Turn,
-      positioning: ShipPositioning
-  ): GameState =
-    gameState.selectedShip match
-      case None =>
-        selectShipAt(gameState, position, turn, positioning)
-      case Some(ship) =>
-        moveSelectedShip(gameState, ship, position, turn, positioning)
-
-  /** Handles cell double click to rotate ship during positioning phase
-    * @param gameState current game state
-    * @param position clicked position
-    * @param turn current turn
-    * @param positioning ship positioning logic
-    * @return updated game state
-    */
-  def handlePositioningDoubleClick(
-      gameState: GameState,
-      position: Position,
-      turn: Turn,
-      positioning: ShipPositioning
-  ): GameState =
-    gameState.selectedShip match
-      case Some(ship) =>
-        rotateSelectedShip(gameState, ship, turn, positioning)
-      case None =>
-        val stateWithSelection = selectShipAt(gameState, position, turn, positioning)
-        stateWithSelection.selectedShip match
-          case Some(selectedShip) =>
-            rotateSelectedShip(stateWithSelection, selectedShip, turn, positioning)
-          case None =>
-            gameState
-
-  /** Handles keyboard input for randomizing ship positions
-    * @param gameState current game state
-    * @param ships list of ships to randomize
-    * @param turn current turn
-    * @param positioning ship positioning logic
-    * @return updated game state
-    */
-  def handleRandomizePositions(
-      gameState: GameState,
-      ships: List[Ship],
-      turn: Turn,
-      positioning: ShipPositioning
-  ): GameState =
-    positioning.randomPositioning(PlayerBoard(), ships) match
-      case Right(newBoard) =>
-        updateBoardForTurn(gameState, newBoard, turn).copy(selectedShip = None)
-      case Left(_) =>
-        gameState
 
   /** Selects a ship at the given position
     * @param gameState current game state
@@ -149,3 +89,65 @@ object PositioningController:
     turn match
       case Turn.FirstPlayer  => gameState.copy(board = newBoard)
       case Turn.SecondPlayer => gameState.copy(enemyBoard = newBoard)
+
+  /** Handles cell click during positioning phase
+    * @param gameState current game state
+    * @param position clicked position
+    * @param turn current turn
+    * @param positioning ship positioning logic
+    * @return updated game state
+    */
+  def handlePositioningClick(
+      gameState: GameState,
+      position: Position,
+      turn: Turn,
+      positioning: ShipPositioning
+  ): GameState =
+    gameState.selectedShip match
+      case None =>
+        selectShipAt(gameState, position, turn, positioning)
+      case Some(ship) =>
+        moveSelectedShip(gameState, ship, position, turn, positioning)
+
+  /** Handles cell double click to rotate ship during positioning phase
+    * @param gameState current game state
+    * @param position clicked position
+    * @param turn current turn
+    * @param positioning ship positioning logic
+    * @return updated game state
+    */
+  def handlePositioningDoubleClick(
+      gameState: GameState,
+      position: Position,
+      turn: Turn,
+      positioning: ShipPositioning
+  ): GameState =
+    gameState.selectedShip match
+      case Some(ship) =>
+        rotateSelectedShip(gameState, ship, turn, positioning)
+      case None =>
+        val stateWithSelection = selectShipAt(gameState, position, turn, positioning)
+        stateWithSelection.selectedShip match
+          case Some(selectedShip) =>
+            rotateSelectedShip(stateWithSelection, selectedShip, turn, positioning)
+          case None =>
+            gameState
+
+  /** Handles keyboard input for randomizing ship positions
+    * @param gameState current game state
+    * @param ships list of ships to randomize
+    * @param turn current turn
+    * @param positioning ship positioning logic
+    * @return updated game state
+    */
+  def handleRandomizePositions(
+      gameState: GameState,
+      ships: List[Ship],
+      turn: Turn,
+      positioning: ShipPositioning
+  ): GameState =
+    positioning.randomPositioning(PlayerBoard(), ships) match
+      case Right(newBoard) =>
+        updateBoardForTurn(gameState, newBoard, turn).copy(selectedShip = None)
+      case Left(_) =>
+        gameState
